+++ client/cody-shared/src/chat/markdown.ts
@@ -2,6 +2,14 @@ import { marked } from 'marked'
 
 import { registerHighlightContributions, renderMarkdown as renderMarkdownCommon } from '@sourcegraph/common'
 
+/**
+ * Supported URIs to render as links in outputted markdown.
+ * - https?: Web
+ * - vscode: VS Code URL scheme (open in editor)
+ * - command:cody.welcome: VS Code command scheme exception we add to support directly linking to the welcome guide from within the chat.
+ */
+const ALLOWED_URI_REGEXP = /^((https?|vscode):\/\/[^\s#$./?].\S*|command:cody.welcome)$/i
+
 const DOMPURIFY_CONFIG = {
     ALLOWED_TAGS: [
         'p',
@@ -35,7 +43,7 @@ const DOMPURIFY_CONFIG = {
         's',
         'u',
     ],
-    ALLOWED_URI_REGEXP: /^(https?|vscode):\/\/[^\s#$./?].\S*$/i,
+    ALLOWED_URI_REGEXP,
 }
 
 /**