+++ client/cody/test/e2e/inline-assist.test.ts
@@ -33,12 +33,4 @@ test('start a fixup job from inline chat with valid auth', async ({ page, sideba
     // Ensures Code Lens is added
     await expect(page.getByText('Edited by Cody')).toBeVisible()
     await expect(page.getByText('<title>Goodbye Cody</title>')).toBeVisible()
-
-    // Ensures Cody's fixup is displayed in comment thread
-    await expect(page.getByText('Check your document for updates from Cody.')).toBeVisible()
-
-    // Ensures Decorations is displayed by checking hover text
-    await page.getByText('>Goodbye Cody<').hover()
-    // The decoration text on hover should start with 'Cody Fixup #' and end with random number
-    await page.getByRole('tooltip', { name: /Cody Assist.*/ }).click()
 })