+++ client/cody/webviews/Login.story.tsx
@@ -43,12 +43,25 @@ const requiresVerifiedEmailAuthStatus: AuthStatus = {
     endpoint,
 }
 
+const supportedAppOS = 'darwin'
+const supportedAppArch = 'arm64'
+const unsupportedAppOS = 'other-os'
+const unsupportedAppArch = 'other-arch'
+
 export default meta
 
 export const Simple: ComponentStoryObj<typeof Login> = {
     render: () => (
         <div style={{ background: 'rgb(28, 33, 40)' }}>
-            <Login authStatus={validAuthStatus} isAppInstalled={false} vscodeAPI={vscodeAPI} setEndpoint={() => {}} />
+            <Login
+                authStatus={validAuthStatus}
+                isAppInstalled={false}
+                vscodeAPI={vscodeAPI}
+                onLoginRedirect={() => {}}
+                endpoint={endpoint}
+                appOS={supportedAppOS}
+                appArch={supportedAppArch}
+            />
         </div>
     ),
 }
@@ -60,7 +73,10 @@ export const InvalidLogin: ComponentStoryObj<typeof Login> = {
                 authStatus={invalidAccessTokenAuthStatus}
                 isAppInstalled={false}
                 vscodeAPI={vscodeAPI}
-                setEndpoint={() => {}}
+                onLoginRedirect={() => {}}
+                endpoint={endpoint}
+                appOS={supportedAppOS}
+                appArch={supportedAppArch}
             />
         </div>
     ),
@@ -73,7 +89,10 @@ export const UnverifiedEmailLogin: ComponentStoryObj<typeof Login> = {
                 authStatus={requiresVerifiedEmailAuthStatus}
                 isAppInstalled={false}
                 vscodeAPI={vscodeAPI}
-                setEndpoint={() => {}}
+                onLoginRedirect={() => {}}
+                endpoint={endpoint}
+                appOS={supportedAppOS}
+                appArch={supportedAppArch}
             />
         </div>
     ),
@@ -82,7 +101,31 @@ export const UnverifiedEmailLogin: ComponentStoryObj<typeof Login> = {
 export const LoginWithAppInstalled: ComponentStoryObj<typeof Login> = {
     render: () => (
         <div style={{ background: 'rgb(28, 33, 40)' }}>
-            <Login authStatus={validAuthStatus} isAppInstalled={true} vscodeAPI={vscodeAPI} setEndpoint={() => {}} />
+            <Login
+                authStatus={validAuthStatus}
+                isAppInstalled={true}
+                vscodeAPI={vscodeAPI}
+                onLoginRedirect={() => {}}
+                endpoint={endpoint}
+                appOS={supportedAppOS}
+                appArch={supportedAppArch}
+            />
+        </div>
+    ),
+}
+
+export const UnsupportedAppOS: ComponentStoryObj<typeof Login> = {
+    render: () => (
+        <div style={{ background: 'rgb(28, 33, 40)' }}>
+            <Login
+                authStatus={validAuthStatus}
+                isAppInstalled={false}
+                vscodeAPI={vscodeAPI}
+                onLoginRedirect={() => {}}
+                endpoint={endpoint}
+                appOS={unsupportedAppOS}
+                appArch={unsupportedAppArch}
+            />
         </div>
     ),
 }