+++ client/cody/webviews/Login.tsx
@@ -3,7 +3,7 @@ import { useCallback } from 'react'
 import { VSCodeButton } from '@vscode/webview-ui-toolkit/react'
 import classNames from 'classnames'
 
-import { AuthStatus, DOTCOM_CALLBACK_URL, DOTCOM_URL, LOCAL_APP_URL } from '../src/chat/protocol'
+import { AuthStatus, DOTCOM_URL, LOCAL_APP_URL } from '../src/chat/protocol'
 
 import { ConnectApp } from './ConnectApp'
 import { ErrorContainer } from './Error'
@@ -13,21 +13,20 @@ import styles from './Login.module.css'
 
 interface LoginProps {
     authStatus?: AuthStatus
-    endpoint?: string
+    endpoint: string | null
     isAppInstalled: boolean
     isAppRunning?: boolean
     vscodeAPI: VSCodeWrapper
     callbackScheme?: string
     appOS?: string
     appArch?: string
-    isAppConnectEnabled?: boolean
-    setEndpoint: (endpoint: string) => void
+    onLoginRedirect: (uri: string) => void
 }
 
 const APP_DESC = {
     getStarted: 'Cody for VS Code requires the Cody desktop app to enable context fetching for your private code.',
     download: 'Download and run the Cody desktop app to configure your local code graph.',
-    connectApp: 'Cody App detected. All that’s left is to do is connect VS Code with Cody App.',
+    connectApp: 'Cody App detected. All that’s left to do is connect VS Code with Cody App.',
     notRunning: 'Cody for VS Code requires the Cody desktop app to enable context fetching for your private code.',
     comingSoon:
         'We’re working on bringing Cody App to your platform. In the meantime, you can try Cody with open source repositories by signing in to Sourcegraph.com.',
@@ -42,17 +41,9 @@ export const Login: React.FunctionComponent<React.PropsWithChildren<LoginProps>>
     appArch,
     isAppInstalled = false,
     isAppRunning = false,
-    isAppConnectEnabled = false,
-    setEndpoint,
+    onLoginRedirect,
 }) => {
     const isOSSupported = appOS === 'darwin' && appArch === 'arm64'
-
-    const loginWithDotCom = useCallback(() => {
-        const callbackUri = new URL(DOTCOM_CALLBACK_URL.href)
-        callbackUri.searchParams.append('requestFrom', callbackScheme === 'vscode-insiders' ? 'CODY_INSIDERS' : 'CODY')
-        setEndpoint(DOTCOM_URL.href)
-        vscodeAPI.postMessage({ command: 'links', value: callbackUri.href })
-    }, [callbackScheme, setEndpoint, vscodeAPI])
 
     const onFooterButtonClick = useCallback(
         (title: 'signin' | 'support') => {
@@ -90,36 +81,8 @@ export const Login: React.FunctionComponent<React.PropsWithChildren<LoginProps>>
         <section className={classNames(styles.section, styles.codyGradient)}>
             <h2 className={styles.sectionHeader}>Cody App for {appOS} coming soon</h2>
             <p className={styles.openMessage}>{APP_DESC.comingSoon}</p>
-            <VSCodeButton className={styles.button} type="button" onClick={() => loginWithDotCom()}>
+            <VSCodeButton className={styles.button} type="button" onClick={() => onLoginRedirect(DOTCOM_URL.href)}>
                 Signin with Sourcegraph.com
-            </VSCodeButton>
-        </section>
-    )
-
-    const EnterpriseSignin: React.FunctionComponent = () => (
-        <section
-            className={classNames(styles.section, !isAppConnectEnabled ? styles.codyGradient : styles.greyGradient)}
-        >
-            <h2 className={styles.sectionHeader}>Sourcegraph Enterprise</h2>
-            <p className={styles.openMessage}>
-                Sign in by entering an access token created through your user settings on Sourcegraph.
-            </p>
-            <VSCodeButton className={styles.button} type="button" onClick={() => onFooterButtonClick('signin')}>
-                Continue with Access Token
-            </VSCodeButton>
-        </section>
-    )
-
-    const DotComSignin: React.FunctionComponent = () => (
-        <section
-            className={classNames(styles.section, !isAppConnectEnabled ? styles.codyGradient : styles.greyGradient)}
-        >
-            <h2 className={styles.sectionHeader}>Sourcegraph.com</h2>
-            <p className={styles.openMessage}>
-                Cody for open source code is available to all users with a Sourcegraph.com account.
-            </p>
-            <VSCodeButton className={styles.button} type="button" onClick={() => loginWithDotCom()}>
-                Continue with Sourcegraph.com
             </VSCodeButton>
         </section>
     )
@@ -133,10 +96,8 @@ export const Login: React.FunctionComponent<React.PropsWithChildren<LoginProps>>
             {authStatus && <ErrorContainer authStatus={authStatus} isApp={isApp} endpoint={endpoint} />}
             {/* Signin Sections */}
             <div className={styles.sectionsContainer}>
-                <EnterpriseSignin />
-                <DotComSignin />
-                {isAppConnectEnabled && <AppConnect />}
-                {isAppConnectEnabled && !isOSSupported && <NoAppConnect />}
+                <AppConnect />
+                {!isOSSupported && <NoAppConnect />}
             </div>
             {/* Footer */}
             <footer className={styles.footer}>