+++ client/web/src/enterprise/app/setup/steps/AppWelcomeSetupStep.tsx
@@ -85,10 +85,10 @@ const EmailCheckVerificationForm: FC = () => {
                 onClick={onNextStep}
             >
                 {hasEmailBeenVerified ? (
-                    <>Email has verified, go to the next step →</>
+                    <>Email verified, go to the next step →</>
                 ) : (
                     <>
-                        <LoadingSpinner /> Waiting for email to be verified
+                        <LoadingSpinner /> Waiting for email verification
                     </>
                 )}
             </Button>