+++ client/jetbrains/CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - Fixed the y position at which autocomplete suggestions are rendered [#53677](https://github.com/sourcegraph/sourcegraph/pull/53677)
 - Fixed rendered completions being cleared after disabling them in settings [#53758](https://github.com/sourcegraph/sourcegraph/pull/53758)
+- Wrap long words in the chat message [#54244](https://github.com/sourcegraph/sourcegraph/pull/54244)
 
 ### Security